requirements.txt
Pillow
numpy
pandas
pip
python-dateutil
pytz
setuptools
six
tornado